bugfix: y must be long for CrossEntropy loss

+++ src/common/lib/models/trainers/trainer_classification.py
@@ -44,7 +44,9 @@ class TrainerClassification(TrainerBase):
         """
         
         logging.info(f"X shape: {X.shape}, y shape: {y.shape}")
-        outputs = self.nova_model.model(X)          
+        outputs = self.nova_model.model(X)      
+        # CrossEntropy loss must get type long for labels
+        y = y.long()    
             
         return {'outputs': outputs, 'targets': y}
